PowerKey Shutter Button (1/2)
Adds an option to make the power key act as the shutter button.

Change-Id: I9d4957601125d0a8ecc1ad3ee0ae285e17df6bd3

Fix force close from PREVENT_POWER_KEY permission being checked even when not requested.

This happens because the permission check occurs without checking the mask. Applications which calls setFlags with a negative flag to remove flags would then force close because the check thinks that the application is requesting the PREVENT_POWER_KEY permission

Change-Id: I645b4f90cd48624ea078eb676fdbf0a766fe13c4