Make LED always behave accordingly to documentation.
http://developer.android.com/reference/android/app/Notification.html#FLAG_SHOW_LIGHTS
says that the notification LED should be lit permanently if onMs > 0 and
offMs == 0. It should be lit off if onMs and offMs both are zero. Both
statements were not enforced in the service so far, leaving it up to the
device specific liblights what to do in those cases.
For consistency (and as no liblights seems to behave accordingly to the
documentation), enforce those statements in the service.